+++ tools/RAiDER/aria/prepFromGUNW.py
@@ -8,7 +8,6 @@
 import os
 from datetime import datetime
 import numpy as np
-import eof.download
 import xarray as xr
 import rasterio
 import pandas as pd
@@ -23,10 +22,10 @@ from RAiDER.logger import logger
 from RAiDER.models import credentials
 from RAiDER.models.hrrr import HRRR_CONUS_COVERAGE_POLYGON, AK_GEO, check_hrrr_dataset_availability
 from RAiDER.s1_azimuth_timing import get_times_for_azimuth_interpolation
-from RAiDER.s1_orbits import ensure_orbit_credentials
+from RAiDER.s1_orbits import download_eofs
 
 ## cube spacing in degrees for each model
-DCT_POSTING = {'HRRR': 0.05, 'HRES': 0.10, 'GMAO': 0.10, 'ERA5': 0.10, 'ERA5T': 0.10}
+DCT_POSTING = {'HRRR': 0.05, 'HRES': 0.10, 'GMAO': 0.10, 'ERA5': 0.10, 'ERA5T': 0.10, 'MERRA2': 0.1}
 
 
 def _get_acq_time_from_gunw_id(gunw_id: str, reference_or_secondary: str) -> datetime:
@@ -95,7 +94,7 @@ def check_weather_model_availability(gunw_path: str,
     ----------
     gunw_path : str
     weather_model_name : str
-        Should be one of 'HRRR', 'HRES', 'ERA5', 'ERA5T', 'GMAO'.
+        Should be one of 'HRRR', 'HRES', 'ERA5', 'ERA5T', 'GMAO', 'MERRA2'.
     Returns
     -------
     bool:
@@ -276,8 +275,7 @@ class GUNW:
         sat = slc.split('_')[0]
         dt  = datetime.strptime(f'{self.dates[0]}T{self.mid_time}', '%Y%m%dT%H:%M:%S')
 
-        ensure_orbit_credentials()
-        path_orb = eof.download.download_eofs([dt], [sat], save_dir=orbit_dir)
+        path_orb = download_eofs([dt], [sat], str(orbit_dir))
 
         return [str(o) for o in path_orb]